feat: Add setting to prevent deletion of used images
This commit introduces a new setting that allows users to prevent the deletion of images that are currently used in at least one note.

A checkbox has been added to the settings panel to enable or disable this feature. The setting is enabled by default. When enabled, attempting to delete a used image from the image management view will result in an alert, and the deletion will be blocked.

The new setting is integrated with both light and dark modes.