fix: Do not show alert when deleting used image
This commit changes the behavior when a user tries to delete an image that is currently in use by a note.

Previously, an alert message was displayed, interrupting the user's workflow. Now, the deletion will simply fail silently if the "prevent used image deletion" setting is enabled, providing a less intrusive experience.